chore: moved fixtures to conftest
